Extract hooks into packages
