+++ packages/@react-aria/color/package.json
@@ -19,6 +19,8 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.6.2",
+    "@react-aria/interactions": "3.2.0",
+    "@react-aria/utils": "3.2.1",
     "@react-stately/color": "^3.0.0-alpha.1",
     "@react-types/color": "^3.0.0-alpha.1"
   },
