+++ packages/@react-aria/color/src/index.ts
@@ -10,5 +10,4 @@
  * governing permissions and limitations under the License.
  */
 
-// export * from './useX';
-export {};
+export * from './useColorWheel';
