+++ packages/@react-spectrum/color/package.json
@@ -41,7 +41,8 @@
     "@react-stately/color": "^3.0.0-alpha.1",
     "@react-stately/utils": "3.1.0",
     "@react-types/color": "3.0.0-alpha.1",
-    "@react-types/shared": "3.2.0"
+    "@react-types/shared": "3.2.0",
+    "@spectrum-icons/workflow": "3.1.0"
   },
   "devDependencies": {
     "@adobe/spectrum-css-temp": "^3.0.0-alpha.1"
