+++ packages/@react-stately/color/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.6.2",
+    "@react-stately/utils": "3.1.0",
     "@react-types/color": "3.0.0-alpha.1"
   },
   "peerDependencies": {
